feat: Add actual delivery date to orders
Introduce a `deliveryDate` field to the `Order` type and backend. This field will store the actual date an order was delivered.

When an order's status is updated to 'Entregado', the `deliveryDate` will be automatically set to the current date and time. This provides more accurate order tracking and history.

The Order Detail page now conditionally displays the actual delivery date when available, otherwise it shows the estimated delivery date.